+++ app/ui/__init__.py
@@ -1,0 +1,3 @@
+from .app_ui import PCDStreamerUI
+
+__all__ = ["PCDStreamerUI"]
