+++ app/utils/camera_utils.py
@@ -5,9 +5,9 @@ import logging
 import threading
 import open3d as o3d
 try:
-    from utils.data_manager.calibration_data import CalibrationData
+    from app.utils.data_manager.calibration_data import CalibrationData
 except ImportError:
-    from data_manager.calibration_data import CalibrationData
+    from app.utils.data_manager.calibration_data import CalibrationData
 # Configure logging
 
 logging.basicConfig(level=logging.DEBUG, format='%(asctime)s [%(levelname)s] %(message)s')
