+++ app/utils/color_palette.py
