--- ui/__init__.py
@@ -1,3 +1,0 @@
-from .application_ui import PCDStreamerUI
-
-__all__ = ["PCDStreamerUI"]
